Add free spaces UI
