+++ components/proposals/components/permissions/components/AddRolesButton.tsx
@@ -1,0 +1,14 @@
+import Button from 'components/common/Button';
+
+type Props = {
+  onClick?: () => void;
+  disabled?: boolean;
+};
+
+export function AddRolesButton({ disabled, onClick }: Props) {
+  return (
+    <Button disabled={disabled} onClick={onClick} variant='text' color='secondary'>
+      + Add roles
+    </Button>
+  );
+}
