+++ components/settings/roles/components/RolePermissions/RolePermissions.tsx
@@ -11,7 +11,7 @@ import { v4 as uuid } from 'uuid';
 import charmClient from 'charmClient';
 import Button from 'components/common/Button';
 import { PostCategoryRolePermissionRow } from 'components/forum/components/PostCategoryPermissions/components/PostCategoryPermissionRow';
-import { ProposalCategoryRolePermissionRow } from 'components/proposals/components/permissions/ProposalCategoryPermissionRow';
+import { ProposalCategoryRolePermissionRow } from 'components/proposals/components/permissions/components/ProposalCategoryPermissionRow';
 import { useProposalCategories } from 'components/proposals/hooks/useProposalCategories';
 import type { UpgradeContext } from 'components/settings/subscription/UpgradeWrapper';
 import { UpgradeChip, UpgradeWrapper } from 'components/settings/subscription/UpgradeWrapper';
